requirements.txt
fastapi[standard]
sqlalchemy
jinja2
easyocr
itsdangerous
pyjwt
